chore: rename to plugin-meta-extractor
