+++ packages/plugin-meta-extractor/README.md
@@ -1,0 +1,9 @@
+# Grafana / Plugin Meta Extractor
+
+A small tool that can be used to extract meta-information from a plugins source code.
+
+---
+
+## Contributing
+
+We are always grateful for contributions! See [CONTRIBUTING.md](../../CONTRIBUTING.md) for more information.
