+++ packages/plugin-meta-extractor/fixtures/defaultExport/chained.tsx
