+++ packages/plugin-meta-extractor/fixtures/defaultExport/mixed.tsx
