+++ packages/plugin-meta-extractor/fixtures/defaultExport/wrapped.ts
