+++ packages/plugin-meta-extractor/fixtures/namedExport/chained.tsx
