+++ packages/plugin-meta-extractor/fixtures/namedExport/mixed.tsx
