+++ packages/plugin-meta-extractor/fixtures/namedExport/wrapped.ts
