+++ packages/plugin-meta-extractor/jest.config.js
