+++ packages/plugin-meta-extractor/src/index.ts
@@ -1,0 +1,1 @@
+export { extractExtensionPoints } from './meta-extractor.js';
