+++ packages/plugin-meta-extractor/src/meta/extensions.ts
