+++ packages/plugin-meta-extractor/tsconfig.json
