--- packages/pme/src/index.ts
@@ -1,1 +1,0 @@
-export { extractExtensionPoints } from './meta-extractor';
